Add support for VFR clips
Using the `+` operator with mismatched frame rates is an error apparently
(but not for BlankClips ???)

Requires splitting the source clip using self.clips() and uses acsuite's
new and improved f2ts function.

So, timecode support could be done for write_to_xml, but it's a lot harder
than it seems, unless I want to run like vspipe --timecodes on a
one-line script for each clip in the self.clips() dict
(which I don't see being a reasonable solution).
Therefore, using a VFR clip is going to be monumentally slow until
VideoNode.frames() gets an upgrade or accessing frameprops is implemented in
a much more efficient way in VapourSynth's core code
(FrameEval can't be used here either).